[feat] - added doctors and patient edit feature
