+++ package.json
@@ -15,6 +15,7 @@
         "date-fns": "^2.26.0",
         "env-cmd": "^10.1.0",
         "framer-motion": "^4",
+        "lodash": "^4.17.21",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "react-icons": "^4.3.1",
