+++ src/Components/Dashboard/UI/Doctors/Doctor.js
@@ -5,27 +5,11 @@ import {
     Icon,
     useColorModeValue,
     Button,
-    RadioGroup,
-    Radio,
-    Modal,
-    ModalOverlay,
-    useToast,
-    ModalContent,
-    ModalHeader,
-    FormControl,
-    FormLabel,
-    Input,
-    ModalFooter,
     Box,
-    ModalBody,
-    ModalCloseButton,
-    InputGroup,
     Stack,
     SimpleGrid,
     ButtonGroup,
-    InputLeftAddon,
     IconButton,
-    Spinner,
     HStack,
 } from '@chakra-ui/react';
 import { AiFillEdit, AiTwotoneLock } from 'react-icons/ai';
@@ -35,201 +19,16 @@ import { BACK_END_URL } from '../../../../env';
 import Cookies from 'universal-cookie';
 import { useNavigate } from 'react-router';
 import { FiUserPlus } from 'react-icons/fi';
+import { InitialFocus } from './Modal/DocInitialFocus';
+import { DocEditModal } from './Modal/DocEditModal';
 const cookie = new Cookies();
 
-function InitialFocus({ isOpen, setOpenModal, loading, setLoading }) {
-    const toast = useToast();
-    const initialRef = React.useRef();
-    const finalRef = React.useRef();
-    const initialState = {
-        name: '',
-        gender: 'male',
-        mobileNumber: '',
-        specialization: 'MBBS',
-        spec: '',
-        position: '',
-    };
-
-    const [data, setData] = useState(initialState);
-    const handleChange = (e) => {
-        console.log(e.target.name, e.target.value);
-        setData({
-            ...data,
-            [e.target.name]: e.target.value,
-        });
-    };
-
-    console.log(data);
-    const handleSubmit = async (e) => {
-        e.preventDefault();
-        setLoading(true);
-
-        try {
-            const response = await axios.post(`${BACK_END_URL}/doctor/create`, data, {
-                headers: {
-                    authorization: cookie.get('session', {
-                        path: '/',
-                    }),
-                },
-            });
-
-            toast({
-                description: 'Doctor Created',
-                position: 'top-right',
-            });
-            console.log(response);
-        } catch (err) {
-            console.log(err);
-            toast({
-                description: err.response.data?.message || "Couldn't Create",
-                position: 'top-right',
-                status: 'error',
-            });
-        }
-        setOpenModal(false);
-        setData(initialState);
-        setLoading(false);
-    };
-
-    return (
-        <>
-            <Modal
-                initialFocusRef={initialRef}
-                finalFocusRef={finalRef}
-                size="xl"
-                isOpen={isOpen}
-                onClose={() => {
-                    setOpenModal(false);
-                    setData(initialState);
-                }}>
-                <ModalOverlay />
-                <ModalContent>
-                    <form onSubmit={handleSubmit}>
-                        <ModalHeader>Add Doctor</ModalHeader>
-                        <ModalCloseButton />
-                        <ModalBody pb={6}>
-                            <HStack mb={2}>
-                                <FormControl>
-                                    <FormLabel>Name</FormLabel>
-                                    <InputGroup>
-                                        <InputLeftAddon children="Dr." />
-                                        <Input
-                                            required
-                                            placeholder="Name"
-                                            value={data.name}
-                                            name="name"
-                                            onChange={handleChange}
-                                        />
-                                    </InputGroup>
-                                </FormControl>
-                            </HStack>
-                            <HStack mb={2}>
-                                <FormControl as="fieldset">
-                                    <FormLabel as="legend">Gender</FormLabel>
-                                    <RadioGroup defaultValue="male">
-                                        <HStack>
-                                            <Radio
-                                                required
-                                                name="gender"
-                                                onChange={handleChange}
-                                                value="female">
-                                                Female
-                                            </Radio>
-                                            <Radio
-                                                required
-                                                name="gender"
-                                                value="male"
-                                                onChange={handleChange}>
-                                                Male
-                                            </Radio>
-                                        </HStack>
-                                    </RadioGroup>
-
-                                    {/* <FormHelperText>Select only if you're a fan.</FormHelperText> */}
-                                </FormControl>
-                                <FormControl>
-                                    <FormLabel>Specialization</FormLabel>
-                                    <Input
-                                        required
-                                        placeholder="Spec"
-                                        value={data.specialization}
-                                        name="specialization"
-                                        onChange={handleChange}
-                                    />
-                                </FormControl>
-                            </HStack>
-                            <HStack mb={2}>
-                                <FormControl>
-                                    <FormLabel>Awards and Positions</FormLabel>
-                                    <Input
-                                        required
-                                        placeholder="Member at IMA, Former Senior Resident"
-                                        type="text"
-                                        value={data.spec}
-                                        name="spec"
-                                        onChange={handleChange}
-                                    />
-                                </FormControl>
-                            </HStack>
-                            <HStack mb={2}>
-                                <FormControl>
-                                    <FormLabel> Current Postitions </FormLabel>
-                                    <Input
-                                        required
-                                        placeholder="Consultant Gynaecologist, Infertility Specialist & Sonologist"
-                                        type="text"
-                                        value={data.position}
-                                        name="position"
-                                        onChange={handleChange}
-                                    />
-                                </FormControl>
-                            </HStack>
-                            <HStack mb={2}>
-                                <FormControl>
-                                    <FormLabel>Mobile</FormLabel>
-                                    <Input
-                                        required
-                                        placeholder="8969846714"
-                                        type="number"
-                                        value={data.mobileNumber}
-                                        name="mobileNumber"
-                                        onChange={handleChange}
-                                    />
-                                </FormControl>
-                            </HStack>
-                        </ModalBody>
-
-                        <ModalFooter>
-                            <Button
-                                type="submit"
-                                disabled={loading ? true : false}
-                                colorScheme="blue"
-                                mr={3}>
-                                {loading ? <Spinner size="sm" /> : 'Save'}
-                            </Button>
-                            <Button
-                                onClick={() => {
-                                    setOpenModal(false);
-                                    setData(initialState);
-                                }}>
-                                Cancel
-                            </Button>
-                        </ModalFooter>
-                    </form>
-                </ModalContent>
-            </Modal>
-        </>
-    );
-}
-
 const Doctor = () => {
-    // const bg = useColorModeValue(
-    //     'white',
-    //     'gray.800'
-    // );
-    // const mobileNav = useDisclosure();
     const [loading, setLoading] = useState(false);
+    const [reload, setReload] = useState(false);
     const [openModal, setOpenModal] = useState(false);
+    const [openEditModal, setOpenEditModal] = useState(false);
+    const [docIndex, setDocIndex] = useState(0);
     const history = useNavigate();
     const bgColor = useColorModeValue('white', 'gray.800');
     const bgColor2 = useColorModeValue('gray.100', 'gray.700');
@@ -253,16 +52,27 @@ const Doctor = () => {
                 console.log(error);
             }
         })();
-    }, [openModal]);
+    }, [reload]);
     return (
         <>
             <Box marginTop="10vh">
-                <InitialFocus
-                    isOpen={openModal}
-                    setOpenModal={setOpenModal}
-                    loading={loading}
-                    setLoading={setLoading}
-                />
+                {openModal && (
+                    <InitialFocus
+                        isOpen={openModal}
+                        setOpenModal={setOpenModal}
+                        loading={loading}
+                        setLoading={setLoading}
+                    />
+                )}
+                {openEditModal && (
+                    <DocEditModal
+                        isOpen={openEditModal}
+                        setOpenModal={setOpenEditModal}
+                        setReload={setReload}
+                        reload={reload}
+                        docDetails={patients[docIndex]}
+                    />
+                )}
                 <Flex alignItems="center" justifyContent="space-between" mx="auto">
                     <HStack display="flex" spacing={3} marginY={5} alignItems="center">
                         <Flex
@@ -379,6 +189,11 @@ const Doctor = () => {
                                                 />
                                                 <IconButton
                                                     colorScheme="green"
+                                                    onClick={() => {
+                                                        console.log('Edit Clicked');
+                                                        setDocIndex(tid);
+                                                        setOpenEditModal(true);
+                                                    }}
                                                     icon={<AiFillEdit />}
                                                 />
                                                 <IconButton
